+++ src/main/java/fr/rhumun/game/worldcraftopengl/content/materials/PurpleLampMaterial.java
@@ -37,10 +37,10 @@ public class PurpleLampMaterial extends PointLight implements InteractableMateri
     @Override
     public void interact(Player player, Block block) {
         if(block.getState() == 0){
-            block.setState(1);
+            block.setState((byte) 1);
         }
         else {
-            block.setState(0);
+            block.setState((byte) 0);
         }
         GAME.getGraphicModule().getLightningsUtils().updateLights();
     }
